refactor(stats): address re-review findings — long subscription count, package-private test seams
- Widen ClientSubscriptionService#getClientSubscriptionsCount() to long (impl
  returns subscriptionCount.sum() instead of intValue()), so the getter matches
  the LongAdder backing, the gauge's sum() read, and the long the historical
  consumer already stores it in — no narrowing on a path built for 100M scale.
- Narrow the three HistoricalStatsTotalConsumer test seams
  (processSaveHistoricalStatsTotal, calculatePairUsingProvidedMsg,
  getTotalMessageCounterPair) from protected to package-private: the test is in
  the same package, there are no subclasses, and package-private expresses
  "test-only, not an extension point" more precisely while keeping the three
  seams consistent.
- Add a drift-guard test asserting a re-subscribe to an already-present filter
  is net-zero (size() - sizeBefore == 0) and does not double-count the total.

+++ application/src/main/java/org/thingsboard/mqtt/broker/actors/client/service/subscription/ClientSubscriptionService.java
@@ -53,5 +53,5 @@ public interface ClientSubscriptionService extends ClientSubscriptionCache {
 
     void clearSubscriptionsInternally(String clientId);
 
-    int getClientSubscriptionsCount();
+    long getClientSubscriptionsCount();
 }
